+++ minigpu_ffi/CHANGELOG.md
@@ -1,6 +1,15 @@
 # minigpu_ffi CHANGELOG
 
-## 1.2.0-WIP
+## 1.2.2-dev-WIP
+
+## 1.2.1
+
+- Fix: fresh builds need dawn find off
+- Change: 1.2.0 migrates minigpu to direct webgpu usage
+- Breaking: Changed setData and references to .write
+- Fix: broken compute shader and buffer finalizers
+
+## 1.2.0
 
 ## 1.1.9