+++ minigpu_platform_interface/CHANGELOG.md
@@ -1,13 +1,24 @@
 # minigpu_platform_interface CHANGELOG
 
-## 1.2.0-WIP
+## 1.2.2-dev-WIP
+
+## 1.2.1
+
+- Fix: fresh builds need dawn find off
+- Change: 1.2.0 migrates minigpu to direct webgpu usage
+- Breaking: Changed setData and references to .write
+- Fix: broken compute shader and buffer finalizers
+
+## 1.2.0
 
 ## 1.1.9
 
 - fix pubspec version issue
+
 ## 1.1.8
 
 - fixed concurrent buffer op crash
+
 ## 1.1.6
 
 - fixed problem with audio input capture providing raw data